+++ Krizaljka/Krizaljka.PostgreSql/KrizaljkaIdeaPostgreSql/KrizaljkaIdeaConfigDao.cs
@@ -16,7 +16,7 @@ internal record KrizaljkaIdeaConfigDao(
     int MaxNumOfCompletedTemplates,
     int ThemeTermsCount,
     int OtherTermsCount,
-    int TemplateIdsOnlyCount,
+    long TemplateIdsOnlyCount,
     int TemplateIdsExcludedCount,
     long CreatedById,
     DateTimeOffset CreatedOn,
