+++ Krizaljka/Krizaljka.WebApi/Models/KrizaljkaIdea/KrizaljkaIdeaConfigResponse.cs
@@ -14,7 +14,7 @@ public record KrizaljkaIdeaConfigResponse(
     int MaxNumOfCompletedTemplates,
     int ThemeTermsCount,
     int OtherTermsCount,
-    int TemplateIdsOnlyCount,
+    long TemplateIdsOnlyCount,
     int TemplateIdsExcludedCount,
     long CreatedById,
     DateTimeOffset CreatedOn,
